fix: reminderid type을 long type으로 수정

+++ doormate/src/main/java/com/example/doormate/domain/Reminder.java
@@ -14,7 +14,7 @@ public class Reminder {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
-    private int reminderId;
+    private long reminderId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User userId;
